refactor: 유저 서비스 DTO 패키지 변경
