+++ src/main/java/coffeemeet/server/user/controller/UserController.java
@@ -7,10 +7,10 @@ import coffeemeet.server.user.controller.dto.MyProfileHttpDto;
 import coffeemeet.server.user.controller.dto.UserProfileHttpDto;
 import coffeemeet.server.user.domain.OAuthProvider;
 import coffeemeet.server.user.controller.dto.AuthInfo;
-import coffeemeet.server.user.dto.MyProfileDto;
+import coffeemeet.server.user.service.dto.MyProfileDto;
 import coffeemeet.server.user.controller.dto.SignupHttpDto;
 import coffeemeet.server.user.controller.dto.UpdateProfileHttpDto;
-import coffeemeet.server.user.dto.UserProfileDto.Response;
+import coffeemeet.server.user.service.dto.UserProfileDto.Response;
 import coffeemeet.server.user.service.UserService;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotBlank;
