+++ src/main/java/coffeemeet/server/user/service/UserService.java
@@ -10,7 +10,6 @@ import coffeemeet.server.common.media.S3MediaService;
 import coffeemeet.server.interest.domain.Keyword;
 import coffeemeet.server.interest.service.cq.InterestCommand;
 import coffeemeet.server.interest.service.cq.InterestQuery;
-import coffeemeet.server.oauth.dto.OAuthUserInfoDto.Response;
 import coffeemeet.server.oauth.service.OAuthService;
 import coffeemeet.server.user.domain.Birth;
 import coffeemeet.server.user.domain.Email;
@@ -18,8 +17,9 @@ import coffeemeet.server.user.domain.OAuthInfo;
 import coffeemeet.server.user.domain.OAuthProvider;
 import coffeemeet.server.user.domain.Profile;
 import coffeemeet.server.user.domain.User;
-import coffeemeet.server.user.dto.MyProfileDto;
-import coffeemeet.server.user.dto.UserProfileDto;
+import coffeemeet.server.user.service.dto.MyProfileDto;
+import coffeemeet.server.user.service.dto.OAuthUserInfo;
+import coffeemeet.server.user.service.dto.UserProfileDto;
 import coffeemeet.server.user.service.cq.UserCommand;
 import coffeemeet.server.user.service.cq.UserQuery;
 import java.io.File;
