--- src/test/java/coffeemeet/server/common/fixture/dto/OAuthInfoResponseFixture.java
@@ -1,13 +1,0 @@
-package coffeemeet.server.common.fixture.dto;
-
-import coffeemeet.server.oauth.dto.OAuthUserInfoDto;
-import org.instancio.Instancio;
-
-public class OAuthInfoResponseFixture {
-
-  public static OAuthUserInfoDto.Response oAuthInfoResponse() {
-    return Instancio.of(OAuthUserInfoDto.Response.class)
-        .create();
-  }
-
-}
