+++ src/test/java/coffeemeet/server/common/fixture/dto/UserProfileDtoFixture.java
@@ -1,6 +1,6 @@
 package coffeemeet.server.common.fixture.dto;
 
-import coffeemeet.server.user.dto.UserProfileDto;
+import coffeemeet.server.user.service.dto.UserProfileDto;
 import org.instancio.Instancio;
 
 public class UserProfileDtoFixture {
