+++ src/test/java/coffeemeet/server/user/controller/UserControllerTest.java
@@ -38,10 +38,10 @@ import coffeemeet.server.common.fixture.dto.UpdateProfileDtoFixture;
 import coffeemeet.server.common.fixture.dto.UserProfileDtoFixture;
 import coffeemeet.server.user.domain.OAuthProvider;
 import coffeemeet.server.user.domain.User;
-import coffeemeet.server.user.dto.MyProfileDto.Response;
+import coffeemeet.server.user.service.dto.MyProfileDto.Response;
 import coffeemeet.server.user.controller.dto.SignupHttpDto;
 import coffeemeet.server.user.controller.dto.UpdateProfileHttpDto.Request;
-import coffeemeet.server.user.dto.UserProfileDto;
+import coffeemeet.server.user.service.dto.UserProfileDto;
 import com.epages.restdocs.apispec.Schema;
 import java.util.Optional;
 import org.junit.jupiter.api.DisplayName;
