+++ src/test/java/coffeemeet/server/user/service/UserServiceTest.java
@@ -36,9 +36,10 @@ import coffeemeet.server.user.domain.Email;
 import coffeemeet.server.user.domain.OAuthInfo;
 import coffeemeet.server.user.domain.Profile;
 import coffeemeet.server.user.domain.User;
-import coffeemeet.server.user.dto.MyProfileDto;
+import coffeemeet.server.user.service.dto.MyProfileDto;
 import coffeemeet.server.user.controller.dto.SignupHttpDto.Request;
-import coffeemeet.server.user.dto.UserProfileDto.Response;
+import coffeemeet.server.user.service.dto.OAuthUserInfo;
+import coffeemeet.server.user.service.dto.UserProfileDto.Response;
 import coffeemeet.server.user.service.cq.UserCommand;
 import coffeemeet.server.user.service.cq.UserQuery;
 import java.io.File;
